backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary,async]==3.2.3
psycopg-pool==3.2.4
openai>=1.57.0
langgraph==0.2.50
pydantic==2.9.2
python-multipart==0.0.12
python-dotenv==1.0.1
